+++ mingw-w64-crt/gdtoa/g_xfmt.c
@@ -69,7 +69,7 @@ char *__g_xfmt (char *buf, void *V, int ndig, size_t bufsize)
 
 	if (ndig < 0)
 		ndig = 0;
-	if (bufsize < ndig + 10)
+	if ((int) bufsize < ndig + 10)
 		return 0;
 
 	L = (UShort *)V;
