+++ mingw-w64-crt/gdtoa/misc.c
@@ -140,7 +140,7 @@ Bigint *Balloc (int k)
 #else
 		len = (sizeof(Bigint) + (x-1)*sizeof(ULong) + sizeof(double) - 1)
 			/sizeof(double);
-		if (k <= Kmax && pmem_next - private_mem + len <= PRIVATE_mem) {
+		if (k <= Kmax && pmem_next - private_mem + len <= (size_t) PRIVATE_mem) {
 			rv = (Bigint*)pmem_next;
 			pmem_next += len;
 		}
