+++ mingw-w64-crt/misc/wctype.c
@@ -58,7 +58,7 @@ static const struct {
 wctype_t wctype (const char *property)
 {
   int i;
-  for (i = 0; i < NCMAP; ++i)
+  for (i = 0; i < (int) NCMAP; ++i)
     if (strcmp (property, cmap[i].name) == 0)
       return cmap[i].flags;
   return 0;
